+++ fearless-utils/src/main/java/jp/co/soramitsu/fearless_utils/encrypt/SignatureVerifierTest.kt
@@ -1,0 +1,5 @@
+package jp.co.soramitsu.fearless_utils.encrypt
+
+class SignatureVerifierTest {
+
+}
